synthetic-review-generator/requirements.txt
# LLM APIs
openai==1.12.0
anthropic==0.18.1
google-genai>=0.2.0  # NEW SDK!

transformers==4.37.2
textblob==0.17.1


# Data handling
pandas==2.2.0
pyyaml==6.0.1

# Utils
python-dotenv==1.0.0
tqdm==4.66.1

# Sentiment analysis
textblob==0.17.1

# Math/Stats
numpy==1.26.3
scikit-learn==1.4.0

# Rate limiting & retry
tenacity==8.2.3

# CLI (optional)
click==8.1.7